moved the bwa mem index image creator to tools package from utils package
